v0.4

+++ Theater-Service/src/main/java/com/example/demo/repositories/TheaterRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.models.Theater;
 
 public interface TheaterRepository extends JpaRepository<Theater, Integer> {
-	List<Theater> findByPlace_PinCode(int pinCode);
+	List<Theater> findByPlace_PinCode(long pinCode);
 }
